fixed a warning

+++ mtools/headers/graphics/pixeldrawer.hpp
@@ -976,7 +976,7 @@ namespace mtools
                 if (nb == nbThreads()) return;
                 _deleteAllThread();
                 _vecThread.resize(nb);
-                for (size_t i = 0; i < nb; i++) { _vecThread[i] = new ThreadPixelDrawer<ObjType>(_obj); }
+                for (int i = 0; i < nb; i++) { _vecThread[i] = new ThreadPixelDrawer<ObjType>(_obj); }
                 }
